api/requirements.txt
azure-functions
torch==2.0.1 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.15.2 --index-url https://download.pytorch.org/whl/cpu
timm==0.9.2
opencv-python-headless==4.8.0.74
numpy==1.24.3
pillow==9.5.0